Fix bug in quantile scale with implicit quantiles.
When the number of quantiles is not specified, the quantile scale will return a
unique value in the range for each value in the domain. However, there was an
off-by-one bug because the internal quantiles array needs one more element than
the number of quantiles to include the minimum value in the domain.