One simulation per frame with force-directed layout.
The previous implementation would run multiple simulation iterations per frame,
which is nice because typically rendering is much more expensive than running
the physics simulation. However, this resulted in a jerky appearance as the
simulation appeared to run in "fast forward" with variable speed. This jerkiness
is exacerbated with very small networks that may drift due to error in numerical
integration.

The new implementation only runs one simulation step per frame, which greatly
improves the fluidity of animation. It also significantly lessens the risk of
drift, as any remaining drift should be so slow as to be imperceptible. Also in
this change, tweak the drag force implementation again, and a bit more jsdoc.