Improvements to interactive Dorling cartogram.
Remove checks for easing; the alert is annoying and disruptive, and there was
also a bug where updating the slider would not have an effect. It seems that we
can just ignore the `ease` state, as a quick fix, though it would be better to
have more robust logic for starting and stopping the simulation. (And, for that
matter, initializing the layout before animating it.)