Add pv.Layout.Bullet.
The bullet layout encapsulates Stephen Few's bullet graph design, which is a
type of bar chart. The bullet graph encodes a set of qualitative ranges (such as
good and bad), a set of performance measures (such as actual and forecast
revenue), and a set of comparative markers (such as previous performance). The
layout supports the four standard orientation, though most "left" and "bottom"
are the most common.

This commit also includes a new "bar" shape for dots. This is identical to the
"tick" shape, except the glyph is centered at the dot location. This allows the
bullet layout to define a marker prototype that can used with either a Rule or a
Dot.