Fixes for enter and exit on transition.
We now recompute all properties for marks that are entering and exiting, rather
than only computing those properties that were explicitly defined. This greatly
reduces the number of properties that typically need to be defined on the enter
and exit prototypes.

This commit also includes some fixes to anchors across transitions. Previously,
the anchors stored a local variable as a side-effect of a def. This does not
work with transitions, given that the enter and exit instances are evaluated
independently. Now we only store the anchor targets in the scenegraph, which
avoids state pollution.