Transition fixes.
We now properly handle mark instances that change visibility. Also, correct a
problem with the recent tweak to anchors if the visible property is overridden.
This commit includes an example of interpolating wedges and anchors; note that
we haven't yet implemented interpolation in polar coordinates, but the plan is
to support that by introducing polar mark types.