Center implied position.
Previously, if both positional properties on a given dimension (e.g., left and
right) were undefined, the left and top properties were implicitly 0. Thus, the
implicit default position was the top-left corner of the panel. Now the center
of the panel is used instead. For example, now wedges can be added to panels
without needing to center them explicitly.