Tweak anchor logic, again.
Having properties with side-effects is bad. Previously, the default anchor data
property had a side-effect of wiring up the target instances. This broke if the
data property was overridden, so now we do this as part of build instead. This
commit also eliminates the need for an anchorTarget() method, as the target
attribute is now set correctly when extending from an anchor. I've retained the
method for now but marked it as deprecated.